chore(gui): extracted the EditableName input in common component
Ticket: MEN-8880
